core/requirements.txt
amqp==5.0.6
anyio==3.2.1
asgiref==3.3.2
beautifulsoup4==4.9.3
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
crispy-bootstrap5==0.3.1
dadata==20.7.0
Django==3.2
django-bootstrap4==3.0.1
django-celery-beat==2.2.0
django-cors-headers==3.12.0
django-crispy-forms==1.11.2
django-filter==2.4.0
django-phonenumber-field==5.0.0
django-tables2==2.3.4
django-timezone-field==4.1.2
djangorestframework==3.12.2
djangorestframework-gis==0.17
drf-extra-fields==3.1.1
environs==9.3.2
et-xmlfile==1.1.0
geographiclib==1.50
geopy==2.1.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.6
httpx==0.18.2
idna==2.10
importlib-metadata==3.10.1
kombu==5.0.2
marshmallow==3.11.1
openpyxl==3.0.7
phonenumberslite==8.12.21
Pillow==8.1.2
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
zipp==3.5.0
